requirements.txt
Flask==3.0.0
google-cloud-pubsub==2.18.4
google-auth==2.23.2